Clean and expose get_audio/video_driver_* funcs on OS class
