+++ core/os/os.cpp
@@ -34,6 +34,7 @@
 #include "input.h"
 #include "os/file_access.h"
 #include "project_settings.h"
+#include "servers/audio_server.h"
 #include "version_generated.gen.h"
 
 #include <stdarg.h>
@@ -625,6 +626,34 @@ void OS::center_window() {
 	int x = scr.width / 2 - wnd.width / 2;
 	int y = scr.height / 2 - wnd.height / 2;
 	set_window_position(Vector2(x, y));
+}
+
+int OS::get_video_driver_count() const {
+
+	return 2;
+}
+
+const char *OS::get_video_driver_name(int p_driver) const {
+
+	switch (p_driver) {
+		case VIDEO_DRIVER_GLES2:
+			return "GLES2";
+		case VIDEO_DRIVER_GLES3:
+		default:
+			return "GLES3";
+	}
+}
+
+int OS::get_audio_driver_count() const {
+
+	return AudioDriverManager::get_driver_count();
+}
+
+const char *OS::get_audio_driver_name(int p_driver) const {
+
+	AudioDriver *driver = AudioDriverManager::get_driver(p_driver);
+	ERR_FAIL_COND_V(!driver, "");
+	return AudioDriverManager::get_driver(p_driver)->get_name();
 }
 
 OS::OS() {
