+++ platform/osx/os_osx.mm
@@ -1026,22 +1026,6 @@ void OS_OSX::set_ime_position(const Point2 &p_pos) {
 	im_position = p_pos;
 }
 
-int OS_OSX::get_video_driver_count() const {
-
-	return 2;
-}
-
-const char *OS_OSX::get_video_driver_name(int p_driver) const {
-
-	switch (p_driver) {
-		case VIDEO_DRIVER_GLES2:
-			return "GLES2";
-		case VIDEO_DRIVER_GLES3:
-		default:
-			return "GLES3";
-	}
-}
-
 void OS_OSX::initialize_core() {
 
 	crash_handler.initialize();
@@ -1214,8 +1198,6 @@ Error OS_OSX::initialize(const VideoMode &p_desired, int p_video_driver, int p_a
 		zoomed = true;
 
 	/*** END OSX INITIALIZATION ***/
-
-	AudioDriverManager::add_driver(&audio_driver);
 
 	// only opengl support here...
 	if (p_video_driver == VIDEO_DRIVER_GLES2) {
@@ -2445,6 +2427,8 @@ OS_OSX::OS_OSX() {
 
 		[NSApp sendEvent:event];
 	}
+
+	AudioDriverManager::add_driver(&audio_driver);
 }
 
 bool OS_OSX::_check_internal_feature_support(const String &p_feature) {
